+++ js/dictionary.js
@@ -1,0 +1,17 @@
+import _ from 'lodash';
+
+const commentText = `Всё отлично!
+  В целом всё неплохо. Но не всё.
+  Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.
+  Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.
+  Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.
+  Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!`;
+
+const names = ['Эомер', 'Фродо', 'Арагорн', 'Арвэн', 'Эовин'];
+
+//eslint-disable-next-line
+const messages = commentText.split(new RegExp('[!\.?]+', 'g'))
+  .map((e) => e.trim())
+  .filter((e) => !_.isEmpty(e));
+
+export {names, messages};
