+++ js/settings.js
@@ -1,0 +1,3 @@
+const userCount = 25;
+
+export {userCount};
